integration test for file task finished

+++ baselibs/apiscall/swagger/model_apis_get_task_resp.go
@@ -10,11 +10,11 @@ package swagger
 
 type ApisGetTaskResp struct {
 	BagName string `json:"bagName,omitempty"`
-	CreateTimeMs int32 `json:"createTimeMs,omitempty"`
-	FinishTimeMs int32 `json:"finishTimeMs,omitempty"`
+	CreateTimeMs int64 `json:"createTimeMs,omitempty"`
+	FinishTimeMs int64 `json:"finishTimeMs,omitempty"`
 	NodeId string `json:"nodeId,omitempty"`
 	Priority int32 `json:"priority,omitempty"`
-	ScheduledTimeMs int32 `json:"scheduledTimeMs,omitempty"`
+	ScheduledTimeMs int64 `json:"scheduledTimeMs,omitempty"`
 	ScriptPath string `json:"scriptPath,omitempty"`
 	TaskDisplayName string `json:"taskDisplayName,omitempty"`
 	TaskName string `json:"taskName,omitempty"`
